Feature/rediger pending (#1033)
* Slår sammen typer for Booking og PendingBooking in frontend.

* Støtte for å redigere bookingønsker

* Støtte for å slette bookingønsker

* Refresh av pending bookings når bruker registrerer eller endrer et bookingønske.

* Viser ikke reservasjonsknapp for leilighet når brukeren allerede har et ønske for den liggende inne.

* deploy til test

---------

+++ my-page-api/src/main/kotlin/no/jpro/mypageapi/controller/PendingBookingController.kt
@@ -8,10 +8,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import jakarta.validation.Valid
 import no.jpro.mypageapi.config.RequiresAdmin
-import no.jpro.mypageapi.dto.BookingDTO
-import no.jpro.mypageapi.dto.CreatePendingBookingDTO
-import no.jpro.mypageapi.dto.PendingBookingDTO
-import no.jpro.mypageapi.dto.PendingBookingTrainDTO
+import no.jpro.mypageapi.dto.*
+import no.jpro.mypageapi.entity.PendingBooking
+import no.jpro.mypageapi.entity.User
 import no.jpro.mypageapi.extensions.getSub
 import no.jpro.mypageapi.service.BookingLotteryService
 import no.jpro.mypageapi.service.PendingBookingService
@@ -106,8 +105,26 @@ class PendingBookingController(
         return ResponseEntity("A new booking has been successfully created", HttpStatus.CREATED)
     }
 
-    @DeleteMapping("{pendingBookingID}")
+    @DeleteMapping("admin/{pendingBookingID}")
     @RequiresAdmin
+    @Operation(summary = "Delete the pending booking connected to the pending booking id")
+    @ApiResponse(
+        responseCode = "200",
+        content = [Content(mediaType = "application/json")]
+    )
+    fun adminDeletePendingBooking(
+        token: JwtAuthenticationToken,
+        @PathVariable("pendingBookingID") pendingBookingID: Long,
+    ): ResponseEntity<String> {
+        val pendingBooking = pendingBookingService.getPendingBooking(pendingBookingID)
+        if (pendingBooking === null) {
+            return ResponseEntity(HttpStatus.NOT_FOUND)
+        }
+        pendingBookingService.deletePendingBooking(pendingBookingID)
+        return ResponseEntity.ok("Pending booking with ID $pendingBookingID has been deleted")
+    }
+
+    @DeleteMapping("{pendingBookingID}")
     @Operation(summary = "Delete the pending booking connected to the pending booking id")
     @ApiResponse(
         responseCode = "200",
@@ -121,7 +138,39 @@ class PendingBookingController(
         if (pendingBooking === null) {
             return ResponseEntity(HttpStatus.NOT_FOUND)
         }
+        val user =
+            userService.getUserBySub(token.getSub()) ?: return ResponseEntity.status(HttpStatus.FORBIDDEN).build()
+        if (!userPermittedToEditBooking(pendingBooking, user)) {
+            return ResponseEntity(HttpStatus.FORBIDDEN)
+        }
+
         pendingBookingService.deletePendingBooking(pendingBookingID)
         return ResponseEntity.ok("Pending booking with ID $pendingBookingID has been deleted")
     }
+
+    @PatchMapping("{pendingBookingId}")
+    @Transactional
+    @Operation(summary = "Edit an existing pending booking")
+    fun editPendingBooking(
+        token: JwtAuthenticationToken,
+        @PathVariable("pendingBookingId") pendingBookingId: Long,
+        @Valid @RequestBody editBookingRequest: UpdateBookingDTO,
+    ): ResponseEntity<String> {
+        val bookingToEdit = pendingBookingService.getPendingBooking(pendingBookingId) ?: return ResponseEntity.notFound().build()
+        val user =
+            userService.getUserBySub(token.getSub()) ?: return ResponseEntity.status(HttpStatus.FORBIDDEN).build()
+
+        if (!userPermittedToEditBooking(bookingToEdit, user)) {
+            return ResponseEntity(HttpStatus.FORBIDDEN)
+        }
+        try {
+            pendingBookingService.editPendingBooking(editBookingRequest, bookingToEdit)
+            return ResponseEntity.ok("The booking has been successfully edited")
+        } catch (e: IllegalArgumentException) {
+            val errorMessage = e.message ?: "An error occurred while editing the booking."
+            throw MyPageRestException(HttpStatus.BAD_REQUEST, errorMessage)
+        }
+    }
+
+    private fun userPermittedToEditBooking(booking: PendingBooking, employee: User) = (booking.employee?.id == employee.id)
 }
